refactor(model) : AnswerNumber, UserNumber를 model 디렉터리로 이동
