+++ src/main/java/baseball/Model/AnswerNumber.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.Model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
